Boutons modifier & publier fonctionnels sur toutes les pages

+++ src/Services/FileUploader.php
@@ -21,7 +21,9 @@ class FileUploader
         return unlink($directory . '/' . $filename);
     }
     public function update(string $oldFilename, string $directory, UploadedFile $file, string $newName = '') {
-        $this->delete($oldFilename, $directory);
+        if ($oldFilename != null) {
+            $this->delete($oldFilename, $directory);
+        }
         return $this->upload($file, $directory, $newName);
 
     }
